requirements.txt
PyQt6==6.4.2
pyperclip==1.8.2
Pillow==10.2.0 
PyInstaller==6.5.0